CA-99 refactored helpers and includes to helpers.h
