+++ csr_validator.cc
@@ -1,25 +1,5 @@
-#include <fstream>
-#include <iostream>
-#include <sstream>
-#include <string>
-
-std::string trim(const std::string& str) {
-    std::string whitespace = " \t\r\n"; // Include carriage return and tab
-    size_t first = str.find_first_not_of(whitespace);
-    if (first == std::string::npos) return "";
-    size_t last = str.find_last_not_of(whitespace);
-    return str.substr(first, (last - first + 1));
-}
-
-
-bool isComment(const std::string& line) {
-    // Check if the first non-whitespace character is '#'
-    for (char ch : line) {
-        if (std::isspace(ch)) continue; // Skip whitespace
-        return ch == '#';
-    }
-    return false;
-}
+#include "helpers.h"
+#include "material_validator.h"
 
 bool parseVersion(std::ifstream& file, std::string& versionLine) {
     std::string line;
@@ -50,6 +30,10 @@ int main(int argc, char** argv) {
         std::cerr << "Unsupported version or version line missing: [" << versionLine << "]" << std::endl;
         return 1;
     }
+
+    std::vector<std::string> materials;
+    std::vector<std::string> textures;
+    std::vector<std::string> primitives;
 
     std::cout << "CSR file validation passed." << std::endl;
