+++ helpers.h
@@ -1,0 +1,14 @@
+#ifndef HELPERS
+#define HELPERS
+
+#include <fstream>
+#include <iostream>
+#include <sstream>
+#include <string>
+#include <vector>
+
+std::string trim(const std::string& str);
+
+bool isComment(const std::string& line);
+
+#endif
